Array of structs

+++ PolynomialAprox/ajustePolV1.c
@@ -12,14 +12,14 @@
 /////////////////////////////////////////////////////////////////////////////////////
 
 void montaSL(double **A, double *b, int n, long long int p, double *x, double *y) {
-	for (int i = 0; i < n; ++i) {
-		for (int j = 0; j < n; ++j) {
+	for (long long int i = 0; i < n; ++i) {
+		for (long long int j = 0; j < n; ++j) {
 			A[i][j] = 0.0;
 				for (long long int k = 0; k < p; ++k)
 					A[i][j] += pow(x[k], i+j);
 		}
 	}
-	for (int i = 0; i < n; ++i) {
+	for (long long int i = 0; i < n; ++i) {
 		b[i] = 0.0;
 		for (long long int k = 0; k < p; ++k)
 			b[i] += pow(x[k],i) * y[k];
@@ -27,9 +27,9 @@ void montaSL(double **A, double *b, int n, long long int p, double *x, double *y
 }
 
 void eliminacaoGauss(double **A, double *b, int n) {
-	for (int i = 0; i < n; ++i) {
-		int iMax = i;
-		for (int k = i+1; k < n; ++k)
+	for (long long int i = 0; i < n; ++i) {
+		long long int iMax = i;
+		for (long long int k = i+1; k < n; ++k)
 			if (A[k][i] > A[iMax][i])
 				iMax = k;
 			if (iMax != i) {
@@ -43,10 +43,10 @@ void eliminacaoGauss(double **A, double *b, int n) {
 				b[iMax] = aux;
 			}
 
-			for (int k = i+1; k < n; ++k) {
+			for (long long int k = i+1; k < n; ++k) {
 				double m = A[k][i] / A[i][i];
 				A[k][i]  = 0.0;
-				for (int j = i+1; j < n; ++j)
+				for (long long int j = i+1; j < n; ++j)
 					A[k][j] -= A[i][j]*m;
 				b[k] -= b[i]*m;
 			}
@@ -54,9 +54,9 @@ void eliminacaoGauss(double **A, double *b, int n) {
 }
 
 void retrossubs(double **A, double *b, double *x, int n) {
-	for (int i = n-1; i >= 0; --i) {
+	for (long long int i = n-1; i >= 0; --i) {
 		x[i] = b[i];
-		for (int j = i+1; j < n; ++j)
+		for (long long int j = i+1; j < n; ++j)
 			x[i] -= A[i][j]*x[j];
 		x[i] /= A[i][i];
 	}
@@ -64,7 +64,7 @@ void retrossubs(double **A, double *b, double *x, int n) {
 
 double P(double x, int N, double *alpha) {
 	double Px = alpha[0];
-	for (int i = 1; i <= N; ++i)
+	for (long long int i = 1; i <= N; ++i)
 		Px += alpha[i]*pow(x,i);
 
 	return Px;
@@ -86,7 +86,7 @@ int main() {
 		scanf("%lf %lf", x+i, y+i);
 
 	double **A = (double **) malloc(sizeof(double *)*n);
-	for (int i = 0; i < n; ++i)
+	for (long long int i = 0; i < n; ++i)
 		A[i] = (double *) malloc(sizeof(double)*n);
 
 	double *b = (double *) malloc(sizeof(double)*n);
@@ -104,7 +104,7 @@ int main() {
 	tEG = timestamp() - tEG;
 
 	// Imprime coeficientes
-	for (int i = 0; i < n; ++i)
+	for (long long int i = 0; i < n; ++i)
 		printf("%1.15e ", alpha[i]);
 	puts("");
